pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pynpm-cli"
version = "0.1.0"
description = "An npm-like package manager for Python that uses pip under the hood"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Subhendu Kumar", email = "internship.subhendu@gmail.com"},
]
keywords = ["package-manager", "pip", "venv", "dependency-management"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
test = ["pytest>=7.0"]

[project.scripts]
pynpm = "pynpm.cli:main"

[project.urls]
Homepage = "https://github.com/subh-skd/pynpm-cli.git"
Repository = "https://github.com/subh-skd/pynpm-cli"
Issues = "https://github.com/subh-skd/pynpm-cli/issues"

[tool.setuptools.packages.find]
where = ["src"]
